Minor Changes to Library
Made a few modifications to the library to improve the look. I
hyperlinked the book title, rather than including a details button,
because it looks cleaner. The button can be added by inserting this code
before the delete button:
<a class="btn btn-inverse btn-small" href="#">Details</a>

I also attempted to add the numeric stepper for qty, but the arrows
didn't come in.