Small Change + Temp Fix
I commented out my broken check in javascript, because it was preventing
the remaining dashboard javascript from functioning properly. I also
added in a line to reload the dashboard once a connection request is
sent (already done for cancelling). The reason for this is to remove the
arguments from the URL, so if they refresh after sending a request, they
don't get the modal again to send a request.