Added Book Check In to Library and Fixed Hashes
The library tabs for checking in books now function the same as on the
dashboard.
Also, I added code to both the library and network so that when a user
clicks on a tab, the hash is added to the URL, so if they refresh they
end up back in the same tab.